pyproject.toml
[project]
name = "one-piece"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework>=3.16.0,<4.0.0",
    "django-cors-headers>=4.7.0,<5.0.0",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sentry-sdk[django] (>=2.29.1,<3.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "twilio (>=9.8.4,<10.0.0)",
    "tblib (>=3.2.0,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "boto3>=1.40.65,<2.0.0",
    "channels (>=4.3.1,<5.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "django-admin-autocomplete-filter (>=0.7.1,<0.8.0)",
    "google-genai>=1.67.0,<2.0.0",
    "whitenoise (>=6.12.0,<7.0.0)",
    "psycopg2-binary>=2.9.11",
    "ruff>=0.15.9",
    "djangorestframework-stubs==3.15.1",
    "drf-yasg>=1.21.15",
    "mypy==1.11.2",
    "django-stubs[compatible-mypy]==5.1.0",
    "pre-commit>=4.5.1",
    "drf-spectacular>=0.29.0",
    "hashids>=1.3.1",
    "redis>=7.4.0",
    "numpy>=2.4.4",
    "colorthief>=0.2.1",
    "opencv-python-headless>=4.13.0.92",
    "redis>=7.4.0",
    "solapi>=5.0.3",
    "korcen>=1.0.2",
    "colorama>=0.4.6",
    "playwright>=1.58.0",
    "django-jazzmin>=3.0.4",
    "requests>=2.33.1",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    # boto3 타입 정보 없어서 검사 못하니까 넘어가게하는 패키지 (boto3는 타임 스텁 원래 제공X)
    "boto3-stubs[s3]>=1.42.88",
    "botocore-stubs>=1.42.41",
    "coverage>=7.13.5",
    "pre-commit>=4.5.1",
]


[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint] #내용검사
select = ["E", "F", "I"]
ignore = ["B008", "B904", "E501", "F405"]
# --fix 옵션을 넣었을 때 자동으로 고쳐줄 규칙들
fixable = ["ALL"] # ruff한테 수정 허락해주는 규칙
unfixable = []

[tool.ruff.format] #글씨체, 정렬
quote-style = "double"
indent-style = "space"


[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true
exclude = ["migrations", "manage.py"]

[[tool.mypy.overrides]]
module = ["admin_auto_filters", "admin_auto_filters.*"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.local"

[tool.coverage.run]
omit = [
    "/migrations/",
    "/tests/",
    "/test_.py",
    "/settings/",
    "/manage.py",
    "/urls.py",
    "/wsgi.py",
    "/asgi.py",
    "*testcase.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true


